+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 - Add `GET` endpoint `/v0/blockInfo/{blockHash}` for getting block information for a given block hash.
+- Add `GET` endpoint `/v0/blocksAtHeight/{blockHeight}` for getting block hashes at a given absolute block height.
 
 - Add `GET` endpoint `/v0/consensusInfo` for getting the current consensus status from the node, including last finalized block, best block, protocol version, and timing parameters.